+++ server/src/document/document.service.ts
@@ -1,10 +1,11 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { HttpException, Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Document } from '../entities/document.entity';
 import { CreateDocumentDto } from './document.dto';
 import { Archive } from 'src/entities/archive.entity';
 import { Subscriber } from 'src/entities/subscriber.entity';
+import { TokenService } from 'src/token/token.service';
 
 
 @Injectable()
@@ -16,6 +17,7 @@ export class DocumentService {
     private archiveRepository: Repository<Archive>,
     @InjectRepository(Subscriber)
     private subscriberRepository: Repository<Subscriber>,
+    private readonly tokenService: TokenService
   ) {}
 
   async createDocument(dto: CreateDocumentDto): Promise<Document> {
@@ -54,13 +56,17 @@ export class DocumentService {
     return document;
   }
 
-  async updateDocument(inventoryNumber: string, updateData: Partial<Document>): Promise<Document> {
+  async updateDocument(inventoryNumber: string, updateData: Partial<Document>, token: string): Promise<Document> {
+    const existUser = await this.tokenService.getUserFromToken(token)
+    if (existUser.role == 'СОТРУДНИК') throw new HttpException('у вас нет прав', 403)
     const document = await this.getDocumentByInventoryNumber(inventoryNumber);
     this.documentRepository.merge(document, updateData);
     return this.documentRepository.save(document);
   }
 
-  async deleteDocument(inventoryNumber: string): Promise<void> {
+  async deleteDocument(inventoryNumber: string, token: string): Promise<void> {
+    const existUser = await this.tokenService.getUserFromToken(token)
+    if (existUser.role == 'СОТРУДНИК') throw new HttpException('у вас нет прав', 403)
     const document = await this.getDocumentByInventoryNumber(inventoryNumber);
     await this.documentRepository.remove(document);
   }
