+++ server/src/subscriber/subscriber.service.ts
@@ -1,14 +1,16 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { HttpException, Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Subscriber } from '../entities/subscriber.entity';
 import { createSubscriberDto } from './subscriber.dto';
+import { TokenService } from 'src/token/token.service';
 
 @Injectable()
 export class SubscriberService {
   constructor(
     @InjectRepository(Subscriber)
     private subscriberRepository: Repository<Subscriber>,
+    private readonly tokenService: TokenService
   ) {}
 
   async createSubscriber({name, department, phone, receivedDate}: createSubscriberDto): Promise<createSubscriberDto> {
@@ -28,13 +30,17 @@ export class SubscriberService {
     return subscriber;
   }
   
-  async updateSubscriber(id: number, updateData: Partial<Subscriber>): Promise<Subscriber> {
+  async updateSubscriber(id: number, updateData: Partial<Subscriber>, token: string): Promise<Subscriber> {
+    const existUser = await this.tokenService.getUserFromToken(token)
+    if (existUser.role !== "АДМИН") throw new HttpException("У вас недостаточно прав", 403)
     const subscriber = await this.getSubscriberById(id);
     this.subscriberRepository.merge(subscriber, updateData);
     return this.subscriberRepository.save(subscriber);
   }
 
-  async deleteSubscriber(id: number): Promise<void> {
+  async deleteSubscriber(id: number, token: string): Promise<void> {
+    const existUser = await this.tokenService.getUserFromToken(token)
+    if (existUser.role !== "АДМИН") throw new HttpException("У вас недостаточно прав", 403)
     const subscriber = await this.getSubscriberById(id);
     await this.subscriberRepository.remove(subscriber);
   }
